+++ src/screens/screen_macros.h
@@ -3,3 +3,4 @@ void enterMacros();
 void exitMacros();
 void drawMacrosScreen();
 void handleMacrosTouch(int x, int y);
+extern void requestMacros();  // defined in CNC_Pendant_UI.cpp